Add the required code to the add-and-remove.js file
